Use material-ui instead of boostrap
Using material-ui simplifies the CSS stuff - previously, there were a
bunch of CSS rules that weren't being used and that wasn't obvious.
Instead, with material-ui / emotion, that isn't the case. Also, we
switch to a grid layout which lets us fit more content on the screen,
especially for wider screens.