Update how inflation is calculated
Previously, we'd calculate inflation from January to the most recent
month that we have available. This mean that we'd show prices _above_
the original price for consoles released during the current year.
Instead, we use the same month for the most recent month _and_ the
previous month - which for the current year, are the same value and for
previous years should still give us a reasonable value.