Remove useDelayedValue
It was a neat idea - however, it wasn't a great fit for the use case.
useDelayedValue was just used to debounce updates to the URL and to the
displayed cards. However, using it in this way forced us to re-render
the page in order to trigger the useEffect hook that would actually
update the page. By getting rid of useDelayedValue, we can eliminate
that re-render.