Redesign the ConsoleCard layout
We put the console name at the top - since that seems like a reasonable
place to expect it to be. This makes it clearer what the value is below
the image - the manufacturer since the name is already clearly at the
top and its not clear what anyone would assume the lower down value
would be. We put the year on the same line as the manufactuer to save
some veritcal space - and it seams reasonable that a user will assume
that the number is a year.