Set the content background color on the <body> element
Setting the background color on the body element makes overscroll on Mac
OS look a bit nicer - previously, we just set it on the content area
which meant scrolling outside of it would show a white background.
Additionally, when a user has searched and found no results, the page
might not fill up the entire screen - which would again leave some white
page visible below the footer bar.