Fix console names and allow for multiple names
Previously, each console could only have a single name - which is
problematice for consoles such as the "Nintendo 64" which is commonly
also called the "N64". So, we update the data to allow for multiple
names. When displaying a name, we just pick the first one - for example,
"Nintendo 64". However, when searching we look through all of them - so
"N64" will produce the expected result. We also fix a few typos in the
names and a typo in a Wikipedia link. When in doubt about the "correct"
name, we just go with what Wikipedia is using.