Remove dependency on csv-parse
Given the very limited data we are dealing with, we don't really need
anything better than splitting on commas which we can do without a
dependency.

Doing this work, we also change how he data is loaded - instead of
loading the months by name, we load them by number. This will be useful
in subsequent changes. Since we are using numbers now, we also need a
function to convert that number into a month name for us for display. In
theory we could probaby use the standard Date API to do this - but I
felt it was just easier and more obvious to use a function with a single
switch statement to do it - at least while we're only supporting
English.