Convert to dates to the new YearMonth API
All of our data has month-level granularity. Previously, we'd been
storing the year and month as separate fields. In order to make handling
of year/month pairs more uniform, we convert everything over to using a
YearMonth type that is just a combination of a year and month value.
This will allow us to define additional functions to work with these
values and get good IDE support / linting.