Implement useUrlState()
useUrlState() replaces useFilterState() / useFilterStateSyncer(). It
does more or less the same thing - but is simpler. Its also more generic
since the state type is now a generic parameter as opposed to being hard
coded into the mechanism that updates the URL with state values.

One significant change is that the useUrlState() debounces updates to
the URL - which is valuable since Chrome and Firefox both get unhappy if
you update the URL too much (ie - ~20 updates/sec - which becomes
possible to do once we add in the slider control in a subsequent
change).

Another update is that we get rid of the reducer that was being used to
handle updates to the state. This makes updates a bit less efficient
since it will triggers slightly more components to re-render. However,
it also simplifies the code some and doesn't seem to have a significant
performance impact.