Load InflationData so that it can be passed to the app
Previously we just used the inflation data when calculating the current
data console prices. However, we want to be able to add a slider to
select prices at some date in the past. So, we'll need to be able to
load the inflation data into the app so that it can be used dynamically
at runtime.

As part of making this data separately loadable, we make some other
changes:

* Data is now loaded into a React Context and accessed that way. This
  just makes the code a bit cleaner since we don't have to pass the data
  around (and the data is basically static - so accessing it from
  Context won't trigger additional re-renders).
* As part of the above, we split data.ts into two files - the
  original file that includes the truly static data (ie, console names
  and when they were released) and loadData.ts - which includes the code
  to load information from the filesystem and can only be accessed via
  getStaticProps().
* We also made some changes to how the data is represented -
  significantly we renamed "raw_oriwg_prices" to "orig_prices" and we
  removed fields that contained formatted values (such as strings
  representing prices with dollar signs and commas) and instead format
  them in the relevant React component.