Convert as much as we can into server components
The big change is to render the Console Picture as a Server Component
which we pass into the CardContainer component (which is a Client
Component). I was hopeful that this would reduce the JSON rendered into
the page - but it doesn't appear to have done that yet. Actually, it
seems to be somewhat worse - we also update to only pass the subset of
data needed to the client components instead of all of the data -
including the various image URLs. I think in theory this will reduce
data in the page in the future since this JSON is smaller than the full
data used by the Server Componenets to render the ConsolePicture
components. However, at least for the time being, Next.js seems to
include both JSON blobs in the final page. I'm hoping that a future
Next.js version will fix this (which also implies that I'm hoping I
didn't do something wrong here - but I couldn't find any place that the
full server data JSON was leaking into the Client Components).

This change also gets rid of the useData() hook in favor of passing the
data explicitly. This was largely done to help make it easier to split
the Server and Client components apart.