Don't display consoles as "Not Released" if at the most recent date
The problem is that the Switch 2 would show up as "Not Released" on the
default view. This is true, but unfortunate since it makes it harder to
compare its announced price to that of other consoles. So, as a special
case, when the view is on the most recent date for which we have
inflation data, we show all consoles as if they have been released
already.