Fix sorting for unreleased consoles
Previously, we decided that if a console was going to be released in the
future, we wanted to use its announced release price as its price (at
least if the user was setting the asOf date to the most recent date that
has data available). However, we hadn't implemented that for sorting -
which meant that unreleased consoles would sort to the end of the list,
which was not desired. So, we fix that.